requirements.txt
streamlit
pdfminer.six
pandas
nltk
spacy
plotly
mysql-connector-python
